GCP_pipeline_project/fastq_job/trigger_fn/requirements.txt
google-cloud-run==0.10.7

